# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: whatsapp
    description: WhatsApp Cloud API send operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /whatsapp/send-template:
    post:
      operationId: whatsappSendTemplate
      tags: [whatsapp]
      summary: Send a WhatsApp template message via Meta Cloud API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappSendTemplateRequest"
      responses:
        "200":
          description: Send result (success or upstream-meta failure, both as JSON)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappSendResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappSendResponse"
        "503":
          description: WhatsApp not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappSendResponse"
  /whatsapp/send:
    post:
      operationId: whatsappSendText
      tags: [whatsapp]
      summary: Send a WhatsApp free-form text message via Meta Cloud API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappSendTextRequest"
      responses:
        "200":
          description: Send result (success or upstream-meta failure, both as JSON)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappSendResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappSendResponse"
        "503":
          description: WhatsApp not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappSendResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    WhatsappSendTemplateRequest:
      type: object
      properties:
        to:
          type: string
          description: Recipient phone (digits only, country-code optional; defaults to +91 if missing)
        template:
          type: string
          description: Approved Meta template name
        language:
          type: string
          description: Template language code (e.g. en, en_US)
          default: en
        params:
          type: array
          items:
            type: string
          description: Body parameter values for the template
      required:
        - to
        - template
    WhatsappSendTextRequest:
      type: object
      properties:
        to:
          type: string
        message:
          type: string
      required:
        - to
        - message
    WhatsappSendResponse:
      type: object
      properties:
        ok:
          type: boolean
        messageId:
          type: string
        error:
          type: string
        code:
          type: number
      required:
        - ok