# lib/api-zod/src/generated/types/whatsappSendResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WhatsappSendResponse {
  ok: boolean;
  messageId?: string;
  error?: string;
  code?: number;
}