# lib/api-zod/src/generated/types/whatsappSendTemplateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WhatsappSendTemplateRequest {
  /** Recipient phone (digits only, country-code optional; defaults to +91 if missing) */
  to: string;
  /** Approved Meta template name */
  template: string;
  /** Template language code (e.g. en, en_US) */
  language?: string;
  /** Body parameter values for the template */
  params?: string[];
}